Swing UI

+++ desktop/src/main/java/komputator/desktop/client/ComputationClient.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 public interface ComputationClient {
 
-    void startComputation(UUID id, int n);
+    void startComputation(UUID id, long n);
 
     ComputationResult getComputationResult(UUID id);
 }
